% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AF_combine.R
\name{AF_combine}
\alias{AF_combine}
\title{AF Combination of P-values}
\usage{
AF_combine(p, weight = 1, log = TRUE)
}
\arguments{
\item{p}{P-values to be combined. A matrix with dimenstions K by B.
If an object of p-values from perm.mar.cov() is used, K is the number
of OTUs, and B is the number of permutataion plus 1.
The first column is observed marginal p-values,
and the rest columns are p-values for permutation.}

\item{weight}{Weights given to the p-values.}

\item{log}{Indicator of whether p-values are on the log scale.}
}
\value{
A list containing the following elements:
        \item{pvalue}{P-value of AF statistic for observed sample (the first element)
                      and all permuted samples.}
        \item{which.selected}{The index of features to be selected.}
}
\description{
This function combines p-values using adaptive Fisher method.
}
