% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cum_prop.R
\name{cum_prop}
\alias{cum_prop}
\title{Taxon Proportion for All Non-rooted Nodes in a Phylogenetic Tree}
\usage{
cum_prop(otu.tab, tree)
}
\arguments{
\item{otu.tab}{OTU count table.
Row - n samples, column - OTUs}

\item{tree}{A rooted phylogenetic tree of R class "phylo".}
}
\value{
A list containing the following elements:
        \item{cum}{Taxon proportion for all non-rooted nodes.
                   The nodes are arranged in the same order as the second column of tree$edge.}
        \item{br.len}{Branch length for all non-rooted nodes.}
        \item{cum.labels}{Labels for all non-rooted nodes.
                          If node.label is not provided by the tree of R class "phylo", we assign "Node k" to the kth node. }
}
\description{
Calculate the taxon proportion for all non-rooted nodes
             in a phylogenetic tree, as well as return the labels for those nodes.
             The code is adapted from GUniFrac package contributed by
             Jun Chen (Jun 12, 2015).
}
