% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiAF.R
\name{MiAF}
\alias{MiAF}
\title{Microbiome Adaptive Fisher Test}
\usage{
MiAF(Y, X, tree, cov = NULL, model = c("gaussian", "binomial"), n.perm)
}
\arguments{
\item{Y}{A numeric vector for continuous or binary responses.}

\item{X}{A matrix of the OTU table.
Row - samples, column - OTUs.}

\item{tree}{A rooted phylogenetic tree of R class "phylo".}

\item{cov}{A data frame/matrix for covariate adjustment.
Row - samples, column - covariates.}

\item{model}{"gaussian" is for the linear regression model (continuous);
"binomial" is for the logistic regression model (binary).}

\item{n.perm}{The number of permutations.}
}
\value{
A list containing the following elements:
        \item{UniFrac}{Unweighted UniFrac-like test p-value.}
        \item{wUniFrac}{Weighted UniFrac-like test p-value.}
        \item{UniFrac5}{Generalized UniFrac-like test p-value.}
        \item{tip.abun}{Leaf-nodes-only test p-value.}
        \item{com}{The p-value of MiAF test that combines p-values of the above four tests.}
        \item{select}{A list of selected associated taxa.
                      lower: under-presented taxa based on lower-tail p-values.
                      upper: over-presented taxa based on upper-tail p-values.}
}
\description{
Microbiome Adaptive Fisher Test
}
