% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select.com}
\alias{select.com}
\title{Associated taxa selected by combining p-values of four tests using different abundance representations}
\usage{
select.com(sel.unweight, sel.weight, sel.5, sel.tip, com)
}
\arguments{
\item{sel.unweight}{The list of selected associated taxa by unweighted UniFrac-like test.}

\item{sel.weight}{The list of selected associated taxa by weighted UniFrac-like test.}

\item{sel.5}{The list of selected associated taxa by generalized UniFrac-like test.}

\item{sel.tip}{The list of selected associated taxa by leaf-nodes-only test.}

\item{com}{The index of selected tests using different abundance representations.
1: unweighted UniFrac-like test.
2: weighted UniFrac-like test.
3: generalized UniFrac-like test.
4: leaf-nodes-only test.}
}
\value{
A list of selected associated taxa after combining four using different abundance representations:
        \item{lower}{Under-presented taxa based on lower-tail p-values.}
        \item{upper}{Over-presented taxa based on upper-tail p-values.}
}
\description{
Associated taxa selected by combining p-values of four tests using different abundance representations
}
