% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AF_combine.R
\name{AF_combine}
\alias{AF_combine}
\title{AF Combination of P-values}
\usage{
AF_combine(p, p.perm, weight = 1, log = TRUE)
}
\arguments{
\item{p}{P-values to be combined.}

\item{p.perm}{A permutation set of p-values.
A matrix with dimensions length(p) by B (the number of permutations).}

\item{weight}{Weights given to the p-values.}

\item{log}{Indicator of whether p-values are on the log scale.}
}
\value{
A list containing the following elements:
        \item{pvalue}{P-value of AF statistic.}
        \item{stat}{AF statistic.}
        \item{p.perm}{P-values of AF statistics for all permuted samples.}
        \item{stat.perm}{AF statistics for all permuted samples.}
        \item{n.selected}{The number of features to be selected.}
        \item{which.selected}{The index of features to be selected.}
}
\description{
This function combines p-values using adaptive Fisher method.
}
