% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perm.mar.cov.R
\name{perm.mar.cov}
\alias{perm.mar.cov}
\title{Score Statistics Permutation}
\usage{
perm.mar.cov(Y, X, binary = FALSE, cov = NULL, nperm, seed)
}
\arguments{
\item{Y}{A numeric vector for continuous or binary responses}

\item{X}{The matrix of transformed taxon proportion.}

\item{binary}{The indicator of whether the response is binary.}

\item{cov}{The matrix/data frame of covariates.}

\item{nperm}{The nnumber of permutations.}

\item{seed}{Set seed.}
}
\value{
A list containing the following elements:
        \item{pval.1side1}{Lower-tail p-values of one-sided score tests.}
        \item{weight}{Standard deviation of transformed microbial abundance,
                     which will be used to calculate the weight.}
}
\description{
Score Statistics Permutation
}
